Slightly optimized
I was able to get operations to run ~ 5 times faster. However it still
takes about 1.5 seconds to do 2^1000 and 2^10000 takes an extremely long
time (minutes at least). Next step may be to change the internal number
storage structure from a string to a deque as changing it to a deque
inside the operator+ increased performance.