Optimized to max without algorithm change
Extremely optimized now, finds 2^1000 in under half a second and scales
better than the library I was testing against.  However, for smaller
powers the other library was faster. To make it even faster I would have
the change the algorithm.

How I optimized was instead of having a temporary deque in the operator+
that was found by pushing back each new digit, I summed the number in
place. (As digits were found I replaced them in the original string and
returned that) I also switched to the direct access instead of the at()
method I had implemented. Although not as pretty, it was much faster as
I could use the local variables for length of the numbers.